damage calc + enemies can heal

+++ Assets/Scripts/Entity.cs
@@ -5,7 +5,7 @@ public abstract class Entity : MonoBehaviour
     public string entityName;
     public int maxHealth;
     public int currentHealth;
-    public int baseStrength;
+    public float baseStrength;
 
 
     public virtual void TakeDamage(int damage)
